+++ packages/extended-angular/src/app/app.component.ts
@@ -1,4 +1,5 @@
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
+import { AppLayoutStateService } from './services/app-layout-state.service';
 
 @Component({
   selector: 'app-root',
@@ -7,10 +8,5 @@ import { Component } from '@angular/core';
   standalone: false
 })
 export class AppComponent {
-  public isDrawerOpen = false;
-  public canCloseDrawer = false;
-
-  public onMenuClicked(): void {
-    this.isDrawerOpen = !this.isDrawerOpen;
-  }
+  public readonly appLayoutState = inject(AppLayoutStateService);
 }